+++ VUE2/src/main/java/edu/tufts/osidimpl/repository/sakai/LongValueIterator.java
@@ -36,7 +36,7 @@ implements org.osid.shared.LongValueIterator
     {
         if (iterator.hasNext())
         {
-            return ((Long)iterator.next()).longValue();
+            return (Long) iterator.next();
         }
         else
         {
