+++ VUE2/src/main/java/tufts/vue/PathwayPanel.java
@@ -856,7 +856,7 @@ public class PathwayPanel extends JPanel
                                                boolean cellHasFocus) {
                 //Get the selected index. (The index param isn't
                 //always valid, so just use the value.)
-                int selectedIndex = ((Integer)value).intValue();
+                int selectedIndex = (Integer) value;
 
                 /*if (isSelected) {
                     setBackground(list.getSelectionBackground());
